Divided signup form into smaller components and changed postal code fetch logic
